+++ src/Enum/RfidWriteProtect.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Janisvepris\ZplBuilder\Enum;
+
+enum RfidWriteProtect: string
+{
+    case NotProtected = '0';
+    case WriteProtected = '1';
+}